New doRollower() method for RotatingFileHandler which keeps the file extension of the log files and inserts the counter number before the extension instead of behind it.
New config variables are introduced to control the FileHandler:

logfilesize allows to specify the size of a single logfiles given in kilobytes

logfilecount specifies the number of rotation logfiles. If 0 is given  no logfile will be archived but the current logfile will reused and the previous data are lost. (New!) If logfilecount is -1, there are no rotating logfiles but the files where renamed if the logfile is full. The newest file gets the highest number until some logfiles where deleted.